+++ src/_utils/interfaces.ts
@@ -11,6 +11,8 @@ export interface CommonFieldsProps {
 export interface A11yProps {
   id?: string
   role?: string
+  title?: string
+  tabIndex?: number
   'aria-label'?: string
   'aria-labelledby'?: string
   'aria-describedby'?: string
@@ -20,17 +22,23 @@ export interface A11yProps {
 type A11yKeys = keyof A11yProps
 
 export function pickA11yProps<T extends A11yProps>(source: T): Pick<T, A11yKeys> {
-  const returnValue = {} as Pick<T, A11yKeys>
+  const initialValue = {} as Pick<T, A11yKeys>
   const keys: A11yKeys[] = [
     'id',
     'role',
+    'title',
+    'tabIndex',
     'aria-label',
     'aria-labelledby',
     'aria-describedby',
     'aria-controls',
   ]
-  keys.forEach(k => {
-    returnValue[k] = source[k]
-  })
-  return returnValue
+
+  return keys.reduce(
+    (acc, key) => ({
+      ...acc,
+      [key]: source[key],
+    }),
+    initialValue,
+  )
 }